Now scraper gets urls from main page
