+++ scraper/package.json
@@ -4,8 +4,8 @@
     "description": "",
     "main": "src/index.js",
     "scripts": {
-        "start": "node src/index.js",
-        "debug": "node --inspect-brk=0.0.0.0:9229 src/index.js",
+        "start": "node --unhandled-rejections=strict src/index.js",
+        "debug": "node --inspect-brk=0.0.0.0:9229 --unhandled-rejections=strict src/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "repository": {
